JRDBのパスワードが正常に受け付けられない問題を習性

+++ KmyKeiba/Models/Connection/JrdbDownloaderModel.cs
@@ -92,8 +92,10 @@ namespace KmyKeiba.Models.Connection
       var url = $"http://www.jrdb.com/member/data/Paci/PACI{dateFormat}.lzh";
 
       var path = Constrants.AppDataDir;
-      var lzhFilePath = Path.Combine(path, "jrdbtmp.lzh");
       var lzhDirPath = Path.Combine(path, "jrdbtmp");
+      var cacheDirPath = Path.Combine(path, "jrdbcache");
+      var lzhFilePath = Path.Combine(path, "jrdbtmp.lzh");
+      var cacheFilePath = Path.Combine(cacheDirPath, $"PACI{dateFormat}.lzh");
 
       try
       {
@@ -107,40 +109,47 @@ namespace KmyKeiba.Models.Connection
       // Basic認証するユーザ名とパスワード
       // 後々セキュリティ
 
-      var myweb = new HttpClient();
-      var request = new HttpRequestMessage
+      if (File.Exists(cacheFilePath))
       {
-        Method = HttpMethod.Post,
-        RequestUri = new Uri(url),
-        Headers = { Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Basic", Convert.ToBase64String(Encoding.ASCII.GetBytes(string.Format("{0}:{1}", id, password)))), },
-      };
+        File.Copy(cacheFilePath, lzhFilePath, true);
+      }
+      else
+      {
+        var myweb = new HttpClient();
+        var request = new HttpRequestMessage
+        {
+          Method = HttpMethod.Get,
+          RequestUri = new Uri(url),
+          Headers = { Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Basic", Convert.ToBase64String(Encoding.ASCII.GetBytes(string.Format("{0}:{1}", id, password)))), },
+        };
 
-      var response = await myweb.SendAsync(request);
-      if (response.StatusCode == HttpStatusCode.Unauthorized)
-      {
-        throw new JrdbDownloadException("IDまたはパスワードが違います");
-      }
-      if (response.StatusCode == HttpStatusCode.ServiceUnavailable)
-      {
-        throw new JrdbDownloadException("サーバーが動作していません");
-      }
-      if (response.StatusCode == HttpStatusCode.InternalServerError)
-      {
-        throw new JrdbDownloadException("サーバーが正常に動作していません");
-      }
-      if (!response.IsSuccessStatusCode)
-      {
-        return;
-      }
+        var response = await myweb.SendAsync(request);
+        if (response.StatusCode == HttpStatusCode.Unauthorized)
+        {
+          throw new JrdbDownloadException("IDまたはパスワードが違います");
+        }
+        if (response.StatusCode == HttpStatusCode.ServiceUnavailable)
+        {
+          throw new JrdbDownloadException("サーバーが動作していません");
+        }
+        if (response.StatusCode == HttpStatusCode.InternalServerError)
+        {
+          throw new JrdbDownloadException("サーバーが正常に動作していません");
+        }
+        if (!response.IsSuccessStatusCode)
+        {
+          return;
+        }
 
-      try
-      {
-        var lzh = await response.Content.ReadAsByteArrayAsync();
-        await File.WriteAllBytesAsync(lzhFilePath, lzh);
-      }
-      catch
-      {
-        throw new JrdbDownloadException("データのダウンロードに失敗しました");
+        try
+        {
+          var lzh = await response.Content.ReadAsByteArrayAsync();
+          await File.WriteAllBytesAsync(lzhFilePath, lzh);
+        }
+        catch
+        {
+          throw new JrdbDownloadException("データのダウンロードに失敗しました");
+        }
       }
 
       // LHA解凍
